refactor: move static functions to utils
